hookup select tag with App.list
