+++ src/cubes/ryans_cube.txt
@@ -354,8 +354,8 @@ Tundra
 Hostage Taker
 The Scarab God
 Baleful Strix
-Lim-Dul's Vault
-Connive
+lim-dûl's vault
+Connive // Concoct
 Creeping Tar Pit
 Fetid Pools
 Polluted Delta
